fix unit test, sort interfaces in comm rank order, clean-up
